+++ docs/source/conf.py
@@ -12,8 +12,11 @@
 #
 import os
 import sys
+import django
 
 sys.path.insert(0, os.path.abspath('../../'))
+os.environ['DJANGO_SETTINGS_MODULE'] = 'tests.test_settings'
+django.setup()
 
 # -- Project information -----------------------------------------------------